+++ app/Filament/Resources/StaticPages/Pages/CreateStaticPage.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\StaticPages\Pages;
+
+use App\Filament\Resources\StaticPages\StaticPageResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateStaticPage extends CreateRecord
+{
+    protected static string $resource = StaticPageResource::class;
+}