+++ app/Filament/Resources/StaticPages/Pages/EditStaticPage.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\StaticPages\Pages;
+
+use App\Filament\Resources\StaticPages\StaticPageResource;
+use Filament\Actions\DeleteAction;
+use Filament\Resources\Pages\EditRecord;
+
+class EditStaticPage extends EditRecord
+{
+    protected static string $resource = StaticPageResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+        ];
+    }
+}